Fixes #4 - disallow duplicate job names on client side
This implementation assumes that only one client may connect to one load balancer (and vice versa), and a node may only deal with one load balancer. Duplicate job name detection is NOT implemented on the load balancer or node, as we assume the only point of entry for jobs is the client.